Record audio in browser
This spike represents the end result of efforts to record audio
in the browser using the recordmp3.js library.
Found a fork at
https://github.com/kehers/Recordmp3js.git
which claims to fix the 50% overrun, and fixed a bug in
it that generates invalid wav files in mono.
The end result sometimes works, either (as committed)
to generate wav files, or (changing recordmp3.js.exportWAV to
call exportSamples and removing the early callback
and return in onmessage) to generate mp3; but most of the
time some or all of the resulting recording is silent.