Beginning to get Gecko45 working in Bloom
- tests pass, except for 4 tests of PDF generation,
which depends on an external program not yet updated
- Gecko is much more prone to throw C# exceptions when
JS executed from C# has a problem. Probably a good many
more things to fix here. Tom is working on making Gecko
give more info when reporting these.
- Gecko no longer raises the C# DomClick event when
a mouse click happens in the main window, probably
because the clicks are in IFrames. Won't get far until
Tom fixes this.