BL-5129 Spike to test PdfJs
This is a minimal test of PdfJs. It replaces Acrobat rather than making it an option as the issue calls for. It also assumes that PdfJs is set up in distfiles/pdfjs; this is not yet automated. (It's 402 files, 10M, so we may not just want to check it in.)