WIP clean up BloomDialog Cancel
BloomDialog needs a prop where the caller supplies a function that the dialog can use to request a close when the backdrop is clicked or escape is pressed. Currently this is called onClose, following MUI dialog, and passes a reason that tells whether it was escape or a backdrop click. But (a) the name onClose suggests something that will happen every time the dialog closes (b) sometimes we should instead call onCancel, especially for the Escape key (c) we wish every caller didn't have to consider the reason; but at least one must, the color picker wants to Cancel on escape and OK on backdrop click. This is a partial implementation of one possible fix.