Major rework of BringBookUpToDate and AppearanceSettings (BL-12853)
The original idea was to allow the user to select a non-legacy theme even if the book has incompatible custom CSS. That branched out in to fixing a multitude of problems, both with AppearanceSettings.cs itself, and pre-existing issues which the new system made more serious.
BringBookUpToDate was drastically reorganized and (hopefully) simplified.
A caching scheme was implemented to allow us to simulate updating various CSS and branding image files when we shouldn't write to the book folder. customBookStyles2.css was introduced as the place to put substitute CSS that allows certain known problem customCss to be replaced with a modern theme.